Fix API#75 https://github.com/curt-labs/API/issues/75
In getLuverneModels() we were limiting our query select to the first element in an array, so not all valid models were included. No we return all elements of the array and filter out values that don't match the year and make when we create the return value.

Fixed luverne_test.go TestGetLuverneStyles test case where the model name no longer includes the make in the name. Was 'Ram 1500', now '1500'

Fixed README.md link to DEVELOPER.md
Added README.md Testing heading, it explains goconvey and gives a usage example.